Always check fthird brance for part-2 completion and diff. module for each fn
