+++ data.js
@@ -1,0 +1,18 @@
+const characterData = {
+
+    hero: {
+       elementId: "hero",
+       name: "Wizard",
+       avatar: "images/wizard.png",
+       health: 60,
+       diceCount:3
+   },   
+    monster: {
+       elementId: "monster",
+       name: "Orc",
+       avatar: "images/orc.png",
+       health: 10,
+       diceCount:1
+   }
+}
+export {characterData}
